payment/requirements.txt
Django==1.3.4
MySQL-python==1.2.4
South==0.7.6
argparse==1.2.1
distribute==0.6.34
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
requests==1.1.0
wsgiref==0.1.2
